services/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
neo4j==5.25.0
firebase-admin==6.5.0
langchain>=0.3.7
langchain-community>=0.3.7
langchain-google-genai==2.0.0
google-generativeai==0.7.2
mesa==2.4.0
scipy>=1.17.1
scikit-learn==1.5.0
numpy==2.1.0
httpx==0.27.0
twilio==9.3.0
pydantic[email]==2.9.0
email-validator==2.2.0
python-multipart==0.0.12
python-dotenv==1.0.0
pyyaml>=6.0
# ── NGO multi-tenancy layer ──────────────────────────────────────────────────
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.36
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis>=5.2.0
